requirements.txt
python-telegram-bot==20.7
requests==2.31.0
schedule==1.2.1
openai==1.13.3
python-dotenv==1.0.1
mercadopago
